refactor: co-locate calculator components in calculator subfolder
