+++ src/components/pool-detail/PoolDetail.jsx
@@ -2,7 +2,7 @@ import { useLoaderData, Link } from "react-router-dom"
 import { useEffect, useState } from "react"
 import { TokenInfoBlock } from "./TokenInfoBlock"
 import { PoolCharts } from "./charts/PoolCharts"
-import { RangeCalculator } from "./RangeCalculator"
+import { RangeCalculator } from "./calculator/RangeCalculator"
 
 export default function PoolDetail() {
    const { pool, history } = useLoaderData()
