+++ src/components/pool-detail/calculator/CalculatorStats.jsx
