+++ src/components/pool-detail/calculator/StrategyComparison.jsx
